etl_pipeline/etl/requirements.txt
pandas
requests
sqlalchemy
pymongo
neo4j
python-dotenv
psycopg2-binary
tqdm